[feat](shotgun) Refactor hiero in ultiple classes
